# util: Pass size to ParseHex to assist preallocation