--- a/src/Makefile.bench.include
+++ b/src/Makefile.bench.include
@@ -34,6 +34,7 @@ bench_bench_bitcoin_SOURCES = \
   bench/rpc_blockchain.cpp \
   bench/rpc_mempool.cpp \
   bench/util_time.cpp \
+  bench/util_bases16_64.cpp \
   bench/verify_script.cpp \
   bench/base58.cpp \
   bench/bech32.cpp \