@sanjaykdragon 2020-02-03
ut: i think the compiler would optimize strlen being called twice into once