@sanjaykdragon 2020-02-04
> So added some benchmarks, and it seems to be within the margin of noise except for `ParseHex`.(which is twice as fast for the 32 byte case) so I think I'll rebase to only do the `ParseHex` pre-allocation
> Before:
> 
> ```
> $ ./src/bench/bench_bitcoin -filter="Base64Decode|Base64DecodeVec|HexParse"
> # Benchmark,    evals,   iterations,  total,      min,         max,       median
> Base64Decode,    5,       1000000,   1.77748,  3.49936e-07, 3.60382e-07, 3.55597e-07
> Base64DecodeVec, 5,       1000000,   1.69569,  3.35597e-07, 3.41911e-07, 3.39301e-07
> HexParse,        5,       1000000,   0.599282, 1.1813e-07,  1.2082e-07,  1.19942e-07
> 
> $ ./src/bench/bench_bitcoin -filter="Base64Decode|Base64DecodeVec|HexParse"
> # Benchmark,    evals,   iterations,  total,      min,         max,       median
> Base64Decode,    5,       1000000,   1.84349,  3.6175e-07,  3.79702e-07, 3.65207e-07
> Base64DecodeVec, 5,       1000000,   1.69212,  3.35303e-07, 3.43094e-07, 3.36332e-07
> HexParse,        5,       1000000,   0.592378, 1.17877e-07, 1.19108e-07, 1.18541e-07
> 
> $ ./src/bench/bench_bitcoin -filter="Base64Decode|Base64DecodeVec|HexParse"
> # Benchmark,    evals,   iterations,  total,      min,         max,       median
> Base64Decode,    5,       1000000,   1.84158,  3.59822e-07, 3.83746e-07, 3.67486e-07
> Base64DecodeVec, 5,       1000000,   1.7276,   3.37973e-07, 3.51833e-07, 3.46255e-07
> HexParse,        5,       1000000,   0.614073, 1.18379e-07, 1.26827e-07, 1.22456e-07
> ```
> 
> After:
> 
> ```
> # Benchmark,    evals,   iterations,  total,      min,         max,       median
> Base64Decode,    5,       1000000,   1.81012,  3.55877e-07, 3.68592e-07, 3.63113e-07
> Base64DecodeVec, 5,       1000000,   1.84868,  3.66254e-07, 3.72895e-07, 3.69377e-07
> HexParse,        5,       1000000,   0.281846, 5.55183e-08, 5.71883e-08, 5.63974e-08
> 
> $ ./src/bench/bench_bitcoin -filter="Base64Decode|Base64DecodeVec|HexParse"
> # Benchmark,    evals,   iterations,  total,      min,         max,       median
> Base64Decode,    5,       1000000,   1.82559, 3.58871e-07, 3.68576e-07, 3.65803e-07
> Base64DecodeVec, 5,       1000000,   1.84521, 3.6735e-07,  3.71235e-07, 3.68715e-07
> HexParse,        5,       1000000,   0.29672, 5.62958e-08, 6.46866e-08, 5.86538e-08
> 
> $ ./src/bench/bench_bitcoin -filter="Base64Decode|Base64DecodeVec|HexParse"
> # Benchmark,    evals,   iterations,  total,      min,         max,       median
> Base64Decode,    5,       1000000,   1.89398,  3.7271e-07,  3.83227e-07, 3.78387e-07
> Base64DecodeVec, 5,       1000000,   1.90284,  3.73173e-07, 3.89818e-07, 3.81565e-07
> HexParse,        5,       1000000,   0.288767, 5.70536e-08, 5.92714e-08, 5.73934e-08
> ```

utACK: if these numbers are real and true, this is a nice improvement