@aureleoules 2022-04-03
I've [rebased the PR](https://github.com/aureleoules/bitcoin/compare/31b1c67cf6d74dd655fafe97e36868b3f8752280...c4844b75c062e559f5ad794ca7fa6ede8329deec) but I can't seem to push to @elichai's branch, @elichai could you allow edits from maintainers on this PR please?
The benchmarks now use the the newer framework and I get the following results:

before:  
```
./src/bench/bench_bitcoin -filter="Base64Decode|Base64DecodeVec|HexParse"
|             ns/byte |              byte/s |    err% |        ins/byte |        cyc/byte |    IPC |       bra/byte |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|                3.41 |      293,134,381.67 |    1.6% |           40.79 |            7.84 |  5.202 |           8.49 |    0.3% |      0.01 | `Base64Decode`
|                3.44 |      290,695,259.27 |    4.2% |           40.79 |            7.91 |  5.158 |           8.49 |    0.3% |      0.01 | `Base64DecodeVec`
|                2.84 |      352,503,546.15 |    1.7% |           34.80 |            6.51 |  5.342 |           7.73 |    0.0% |      0.01 | `HexParse`

./src/bench/bench_bitcoin -filter="Base64Decode|Base64DecodeVec|HexParse"
|             ns/byte |              byte/s |    err% |        ins/byte |        cyc/byte |    IPC |       bra/byte |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|                3.34 |      299,227,204.84 |    1.0% |           40.79 |            7.68 |  5.312 |           8.49 |    0.3% |      0.01 | `Base64Decode`
|                3.43 |      291,367,105.76 |    3.6% |           40.79 |            7.89 |  5.171 |           8.49 |    0.3% |      0.01 | `Base64DecodeVec`
|                2.86 |      349,102,377.98 |    1.6% |           34.80 |            6.58 |  5.287 |           7.73 |    0.0% |      0.01 | `HexParse`

./src/bench/bench_bitcoin -filter="Base64Decode|Base64DecodeVec|HexParse"
|             ns/byte |              byte/s |    err% |        ins/byte |        cyc/byte |    IPC |       bra/byte |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|                3.38 |      295,790,338.45 |    3.1% |           40.79 |            7.77 |  5.249 |           8.49 |    0.3% |      0.01 | `Base64Decode`
|                3.41 |      293,097,336.27 |    3.7% |           40.79 |            7.83 |  5.210 |           8.49 |    0.3% |      0.01 | `Base64DecodeVec`
|                2.84 |      351,733,975.75 |    1.9% |           34.80 |            6.53 |  5.333 |           7.73 |    0.0% |      0.01 | `HexParse`
```

after:
```
./src/bench/bench_bitcoin -filter="Base64Decode|Base64DecodeVec|HexParse"
|             ns/byte |              byte/s |    err% |        ins/byte |        cyc/byte |    IPC |       bra/byte |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|                6.80 |      147,026,345.58 |    2.9% |           40.79 |           15.50 |  2.633 |           8.49 |    0.3% |      0.01 | `Base64Decode`
|                6.74 |      148,375,785.13 |    2.2% |           40.79 |           15.45 |  2.640 |           8.49 |    0.3% |      0.01 | `Base64DecodeVec`
|                1.59 |      626,996,251.24 |    3.6% |           15.36 |            3.67 |  4.189 |           3.34 |    0.0% |      0.01 | `HexParse`

./src/bench/bench_bitcoin -filter="Base64Decode|Base64DecodeVec|HexParse"
|             ns/byte |              byte/s |    err% |        ins/byte |        cyc/byte |    IPC |       bra/byte |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|                3.48 |      287,720,445.40 |    3.5% |           40.79 |            7.98 |  5.114 |           8.49 |    0.3% |      0.01 | `Base64Decode`
|                3.65 |      274,213,136.51 |    1.2% |           40.79 |            8.34 |  4.892 |           8.49 |    0.3% |      0.01 | `Base64DecodeVec`
|                1.06 |      945,347,119.65 |    1.5% |           15.36 |            2.39 |  6.438 |           3.34 |    0.0% |      0.01 | `HexParse`

./src/bench/bench_bitcoin -filter="Base64Decode|Base64DecodeVec|HexParse"
|             ns/byte |              byte/s |    err% |        ins/byte |        cyc/byte |    IPC |       bra/byte |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|                3.61 |      277,048,390.25 |    4.4% |           40.79 |            7.96 |  5.128 |           8.49 |    0.3% |      0.01 | `Base64Decode`
|                3.48 |      287,615,390.01 |    3.2% |           40.79 |            7.87 |  5.180 |           8.49 |    0.3% |      0.01 | `Base64DecodeVec`
|                1.04 |      965,354,790.91 |    2.5% |           15.36 |            2.37 |  6.488 |           3.34 |    0.0% |      0.01 | `HexParse`
```